Add Copy and CopyArray functions
The Copy and CopyArray functions provide a way to deep copy a JSON map;
values in such a map which are not one of the JSON types are not
deep-copied. This should be faster than a 'proper' deep-copy, where
reflection is used to recurse into arbitrary pointer types, and should
be sufficient for working with JSON-like data.