Add mpath, errctx, and consterr sub-packages
This adds the mpath, errctx, and consterr sub-packages. These packages
provide a method of using the `Get*` and `Pop*` functions from the main
package in such a way that error handling can be streamed to various
handlers without early exits. This is useful for tools which want to
output all errors encountered instead of stopping once the first error
is encountered.