Add Optional access functions
This adds functions for getting optional values out of maps. These
functions take a default value and return that on error or if the value
was not present in the map.

Additionally, this adds tests for the MissingRequiredValueError type,
since those were forgotten when adding it. This brings test coverage
back up to 100%.