Added unpack package
This adds an `unpack` package which implements functions useful for
unpacking values, where explicit handling of errors when fetching values
is not necessary due to using an error context.

This also refactors the enum checking code into a separate function,
`maputil.CheckEnum`, which allows the `unpack` package to use it.